hive_commons/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "hive-commons"
version = "0.1.0"
description = "MidOS shared infrastructure"
readme = "README.md"
requires-python = ">=3.11"
license = "MIT"
authors = [
    { name = "Sovereign Hive", email = "hive@local" }
]

dependencies = [
    "structlog>=24.0.0",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",
    "lancedb>=0.29.0,<0.30.0",
    "google-genai>=1.0.0",
    "pydantic>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "ruff>=0.5.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src/hive_commons"]

[tool.hatch.build.targets.wheel.sources]
"src" = ""

[tool.hatch.build.targets.editable]
dev-mode-dirs = ["src"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
ignore = ["E501"]
